Add extended query support to QueryHook
Add extended query support to QueryHook,.

Also re-use the parsed statement instead of parsing it again because it
is was wasted work, and it also made the code less clean.

(We still potentially handle the params more than once, which can be
fixed later)

Finally, ensure the client is Send + Sync. (Should this be enforced in
pgwire?)