Describe table 0 (#252)
* Improve `pg_get_constraintdef` UDF

`pg_get_constraintdef` can optionally take a boolean as a second
parameter, so update the UDF to support this.

* Implement `pg_partition_ancestors`

Add a UDF for `pg_partition_ancestors`

* Add tests for describe table in psql

Add test for describing an example table (`\d customer`) to ensure we
handle it properly.  Most of the queries already worked fine (only 2
failed) but I added them all for completeness.